Time Log Grid View
Optimized the time log grid view to use to pair login events with logoff
events on alternating lines. I also added the code that will be used to
automatically write the time log events into the database, but I'm
keeping it commented out for now.

I also added a local source for the jquery library so it will work for
people who are blocked.